backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "This is the full dependencies list for the backend subdirectory."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "azure-ai-formrecognizer>=3.3.3",
    "bcrypt>=5.0.0",
    "docker>=7.1.0",
    "fastapi>=0.133.1",
    "google-genai>=1.63.0",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "opencv-python>=4.13.0.92",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.1.1",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "psycopg2-binary>=2.9.11",
    "ruff>=0.15.1",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]

[tool.black]
target-version = ["py312"]
line-length = 88
